+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 * feat: add status api to detect api ready by @hunjixin /添加状态检测接口 [[#282](https://github.com/ipfs-force-community/droplet/pull/282)]
 * feat: Add miner manager related commands @diwufeiwen /增加 miner manager 相关命令 [[#293](https://github.com/ipfs-force-community/droplet/pull/293)] 
 * feat: add command to print signer deal detail by @simlecode /增加两个命令行用于输出单个存储订单和检索的详情 [[#298](https://github.com/ipfs-force-community/droplet/pull/298)]
-* feat: unsealed from sp through venus-gateway / 通过venus-gateway给SP下发unsealed请求 by @diwufeiwen [[#267](https://github.com/ipfs-force-community/droplet/pull/267)]
+* feat: unsealed from sp through sophon-gateway / 通过sophon-gateway给SP下发unsealed请求 by @diwufeiwen [[#267](https://github.com/ipfs-force-community/droplet/pull/267)]
 * feat: opt deal list cmds by @simlecode / 订单查询优化 [[#301](https://github.com/ipfs-force-community/droplet/pull/301)] 
 * feat: new api ReleaseDeals by @0x5459 /新增 ReleaseDeals 接口 [[#305](https://github.com/ipfs-force-community/droplet/pull/305)]
 * feat: add command to cancel data transfer by @simlecode / 根据条件取消检索订单 [[#307](https://github.com/ipfs-force-community/droplet/pull/307)]
@@ -42,5 +42,5 @@
 
 * 增加列出 storage/retrieval asks 命令行 [[#272](https://github.com/ipfs-force-community/droplet/pull/272)]
 * 重构 updatedealstatus 接口 [[#289](https://github.com/ipfs-force-community/droplet/pull/289)]
-* 升级 venus、venus-messager、venus-gateway 和 venus-auth 版本到 v1.10.0
+* 升级 venus、sophon-messager、sophon-gateway 和 sophon-auth 版本到 v1.10.0
 * 升级 go-jsonrpc 版本到 v0.1.7